Still green on my first, scenario after removing my code for directory handling in favour of Aruba's code. In particular, I'm able to remove a lot of the code to do with PATH management in the cucumber tests since Aruba ensures that the project's bin directory will be on the path for any of the step definitions.
I also took the opportunity to remove a lot of the duplication between
my skeleton implementation and the test. The upshot of this is that I'm
no longer referring to AppA explicitly in the implementation.

Ditched the string formatting report for standard ERB template.

Still have mixed feelings about the commandline arguments that allow for
the selection of the correct report and format. These are purely
cosmetic since there is only one report and one format. So although I'm
using the tests to say up-front that I expect this is how users will
interact with the program, I don't have any competing scnenarios at the
moment which would drive out the need for these (currently ignnored)
commandline args to be interpreted properly.

Next step is to write a slightly more interesting scenario that will
require focus to turn towards decomposition of the problem into unit
tested classes with clean interfaces between them. I've held off doing
this as long as I dare. As you can tell, I don't dare very far. This
first scenario is more about ensuring that I can end-to-end feedback
from my tests as my users would experience the applications.