Introduced the PackageReport class to hold the complexity of the report generation logic currently in LongHardLook. As I drive out the PackageReport through unit tests, there's a necessity to isolate the test from slow, fragile things like the filesystem aswell as limiting the scope of the functionality being tested.
As I do this, I notice that I'm forced to consier all the relationships
my PackageReport will have to these components of the system like the UI
and the filesystem. Since it's my adapter layer, I get to write the code
I wish I had. I use mocks and stubs to drive out the interfaces with
objects like the FilesystemAdapter which currently doesn't exist.

I need to consider what kind of relationship they are:
- dependency (in which case I'd use an initializer arg)
- notification (in which case I might have a callback interface)
- adjustment (in which case I might use a block)

The requirement to be isolated from infrastructure brings to the fore
the fact that I rely upon the filesystem. This doesn't show up in my
cucumber scenario so much, since I'm able to use the real filesystem.

Stepping back at the end of the unit test, it's done the job of
isolating the contribution of the PackageReport and formalizing the
relationships it has with the UI and an as-yet non-existent
FileSystemAdapter. However, it still looks a little odd when I consider
the complexity of the interactions being stubbed and the fact that my
Report should be dealing directly in terms of FileSystems. Still, it's
an incremental improvement. It's not perfect, but we should never let
perfection be the enemy of improvement. This is a transitional state.